python-reports/requirements.txt
fastapi==0.105.0
uvicorn==0.24.0
sqlalchemy==2.0.23
pymysql==1.1.0
psycopg2-binary==2.9.9
python-dotenv==1.0.0
pydantic==2.5.2
httpx==0.25.2 